chromium/third_party/blink/web_tests/web-components-v0-only/fast/dom-custom/imports-block-upgrade-in-client.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
var upgradeIds = [];
</script>
<x-a id='a1'></x-a>
<link rel="import" href="resources/import-block-upgrade-in-client-child.html">
<body>
<x-a id='a2'></x-a>
<script>
test(function () {
    assert_array_equals(window.upgradeIds, ['a1', 'a2'], 'Elements in the client should be upgraded by a definition in an import.');
}, 'upgrade in the client');
</script>
</body>
